Move payment-methods directory to billing-payment-methods
